Work around GLIBC issue for powerpc. Only for USE_CUSTOM_BAUD.
For powerpc the termios ioctl numbers are defined with reference to the
argument-structure size. This structure must be the one expected by the
kernel (kernel termios) not the one defined by GLIBC (in
termios.h). Work around this by defining ioctl numbers with reference
to the privately defined termios2 structure.

Applies only to Linux/PowePC, and only if USE_CUSTOM_BAUD is enabled.