Correct term_find() users to use indirect fd
The functions use their file descriptor argument fd to locate the
term_s structure t, but then rely on the fact that t->fd == fd.
This is not incorrect, but to assist in future refactoring steps,
prefer to use t->fd to access the file descriptor, instead of using
the fd argument directly.