Remove 'add' behaviour from term_set()
term_set() previously included the behaviour of term_add() and
term_erase(). But this makes it harder to change term_add().

This commit removes the add/erase behaviour from term_set().
The single caller of term_set() relyig on the old behaviour
is udpated to use this form:

	term_add(fd);
	r = term_set(fd, ...);
	if ( r < 0 )
		term_erase(fd);

This is preparation work for changing the signature of term_add().