Disable send-/receive-file cmds for RFC2217 ports
Send and receive file commands pass the port filedes to a newly exec'ed
binary, as its stdin/stdout. This obviously will not work for RFC2217
ports (as the data-stream contains inline signaling that must be removed
by the term framework).

In order to support send- and receive- file commands for RFC2217 ports,
the execed binary must be connected (probably through pipes) to a newly
forked process that monitors the pipes and forwards the data to the
RFC2217 port using the term_read() / term_write() functions that do the
necessary filtering. This may still not work if the execed binary
expects to find a tty connected to its stdin/stdout. In this case
there's nothing we can do...

Anyway, this is too much work so, for now, so I just disable the
commands.