Better baudrate parsing/setting for RFC2217 ports
Use the same code to parse the baudrate from a termios structure and to
write the baudrate to a termios structure for both: native
and RFC2217 ports. Also use the same code to translate between
speeds (expressed in bits-per-second) and budrate-codes.

Seems to work ok, but more testing is needed, especially around some
corner-cases.