Fixed bug in mapping-specifiers completion
The completion script tries to remove (not suggest) mappings you have
already specified. This must not happen for the currently under
completion (i.e. last) mapping (if it matches a specifier). Example: If
you try to complete "lfcr" (no blank of comma at the end) then the list
of possible completions must still include "lfcr" along with
"lfcrlf". If not, readline will replace the current word with "lfcrlf",
which is wrong